Overview-Staff-Version: Restructure & stylesheet
