+++ client/src/screens/StaffScreens/EHROverviewDVScreen/styles.js
@@ -1,0 +1,8 @@
+import { StyleSheet } from "react-native";
+import theme from "../../../theme.style";
+
+const styles = StyleSheet.create({
+    
+});
+
+export default styles;
